requirements.txt
redis
lxml
pyelasticsearch
requests
beautifulsoup4
python_dateutil
grequests
python-logstash
SQLAlchemy
tqdm
apscheduler
pytz
begins
rq
